Create boardSchema
